// frontend/src/app/core/api/model/payment-transaction-response.ts
/**
 * Training school system Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PaymentTransactionResponse { 
    id?: number;
    receiptNumber?: string;
    financialContractId?: number;
    contractNumber?: string;
    amount?: number;
    paymentMethod?: PaymentTransactionResponse.PaymentMethodEnum;
    paymentDate?: string;
    reference?: string;
    status?: PaymentTransactionResponse.StatusEnum;
    createdDate?: string;
    modifiedDate?: string;
}
export namespace PaymentTransactionResponse {
    export const PaymentMethodEnum = {
        Cash: 'CASH',
        Check: 'CHECK',
        BankTransfer: 'BANK_TRANSFER',
        CreditCard: 'CREDIT_CARD',
        OnlineStripe: 'ONLINE_STRIPE'
    } as const;
    export type PaymentMethodEnum = typeof PaymentMethodEnum[keyof typeof PaymentMethodEnum];
    export const StatusEnum = {
        Pending: 'PENDING',
        Cleared: 'CLEARED',
        Bounced: 'BOUNCED',
        Cancelled: 'CANCELLED',
        Refunded: 'REFUNDED'
    } as const;
    export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];
}